+++ test/geo/value_geo_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:value_objects_pkg/src/geo/value_geo.dart';
 
 void main() {