+++ test/text/value_text_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:value_objects_pkg/src/core/value_failure.dart';
 import 'package:value_objects_pkg/src/text/value_text.dart';